player: Two int -> int64_t fixes. Should be able to play more than a few hours now without crashing.

+++ common/PEQ_clock.c
@@ -127,8 +127,8 @@ void UpdateClock(
 ){
 	int clock_minutes,clock_seconds;
 	char temp[52];
-	int time=clock_time;
-        static int last_time = -1;
+	int64_t time=clock_time;
+        static int64_t last_time = -1;
 
 	if( ! pc->isplaying){
           if(last_time==-2)
